+++ tests/testthat/test-snapshot.R
@@ -397,13 +397,11 @@ test_that("packages installed from CRAN using pak are handled", {
   library <- renv_paths_library()
   ensure_directory(library)
   pak <- renv_namespace_load("pak")
-  suppressMessages(pak$pkg_install("toast"))
+  quietly(pak$pkg_install("toast"))
   record <- renv_snapshot_description(package = "toast")
 
-  expect_named(
-    record,
-    c("Package", "Version", "Source", "Repository", "Requirements", "Hash")
-  )
+  expected <- c("Package", "Version", "Source", "Repository")
+  expect_contains(names(record), expected)
 
   expect_identical(record$Source, "Repository")
   expect_identical(record$Repository, "CRAN")
@@ -616,12 +614,38 @@ test_that("standard remotes drop RemoteSha if it's a version", {
 
 })
 
-test_that("we can produce old-style lockfiles if requested", {
+test_that("a package's hash can be re-generated from lockfile", {
+  
+  project <- renv_tests_scope("breakfast")
+  init()
+  
+  lockfile <- snapshot(lockfile = NULL)
+  records <- renv_lockfile_records(lockfile)
+  
+  enumerate(records, function(package, record) {
+    path <- system.file("DESCRIPTION", package = package)
+    actual <- renv_hash_description(path)
+    expected <- renv_hash_record(record)
+    expect_equal(actual, expected)
+  })
+  
+})
 
-  skip_on_cran()
+test_that("lockfiles are stable (v1)", {
+  
+  renv_scope_options(renv.lockfile.version = 1L)
+  
+  project <- renv_tests_scope("breakfast")
+  init()
+  
+  expect_snapshot(. <- writeLines(readLines("renv.lock")))
+  
+})
 
-  renv_scope_options(renv.lockfile.minimal = TRUE)
-
+test_that("lockfiles are stable (v2)", {
+  
+  renv_scope_options(renv.lockfile.version = 2L)
+  
   project <- renv_tests_scope("breakfast")
   init()