+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 1.1.0  (UNRELEASED)
 
+* `renv` now includes the contents of each package's DESCRIPTION file in
+  the package records for generated lockfiles. (#2057)
+
 * Fixed an issue where `renv::snapshot()` could fail if invoked within
   a project containing empty or invalid `.ipynb` files. (#2073)